+++ tests/testthat/test-widget-set-value.R
@@ -1,90 +1,47 @@
-
-context("Widget$setValue")
-
-test_that("Widget$setValue for all input widgets", {
-
+test_that("can set values of all input controls", {
   app <- ShinyDriver$new(test_path("apps/081-widgets-gallery"))
 
-  ## numericInput
-  expect_identical(
-    app$findWidget("num")$setValue(42)$getValue(),
-    42L
-  )
+  roundtrip <- function(app, name, value) {
+    app$findWidget(name)$setValue(value)$getValue()
+  }
 
-  ## selectInput
-  expect_identical(
-    app$findWidget("select")$setValue(2)$getValue(),
-    "2"
-  )
+  expect_identical(roundtrip(app, "num", 42), 42L)
+  expect_identical(roundtrip(app, "select", 2), "2")
 
   ## checkboxInput
-  expect_true(
-    app$findWidget("checkbox")$setValue(TRUE)$getValue()
-  )
-  expect_false(
-    app$findWidget("checkbox")$setValue(FALSE)$getValue()
-  )
+  expect_true(roundtrip(app, "checkbox", TRUE))
+  expect_false(roundtrip(app, "checkbox", FALSE))
 
   ## checkboxGroupInput
-  expect_equal(
-    app$findWidget("checkGroup")$setValue(c("1", "2"))$getValue(),
-    c("1", "2")
-  )
-  expect_equal(
-    app$findWidget("checkGroup")$setValue(c("3"))$getValue(),
-    "3"
-  )
-  expect_equal(
-    app$findWidget("checkGroup")$setValue(character())$getValue(),
-    character()
-  )
+  expect_equal(roundtrip(app, "checkGroup", c("1", "2")), c("1", "2"))
+  expect_equal(roundtrip(app, "checkGroup", "3"), "3")
+  expect_equal(roundtrip(app, "checkGroup", character()), character())
 
   ## dateInput
-  expect_equal(
-    app$findWidget("date")$setValue(Sys.Date())$getValue(),
-    Sys.Date()
-  )
-  expect_equal(
-    app$findWidget("date")$setValue(as.Date("2012-06-30"))$getValue(),
-    as.Date("2012-06-30")
-  )
+  date <- as.Date("2012-06-30")
+  expect_equal(roundtrip(app, "date", date), date)
 
   ## dateRangeInput
-  v <- as.Date(c("2012-06-30", "2015-01-21"))
-  expect_equal(
-    as.character(app$findWidget("dates")$setValue(v)$getValue()),
-    as.character(v)
-  )
+  dates <- as.Date(c("2012-06-30", "2015-01-21"))
+  expect_equal(roundtrip(app, "dates", dates), dates)
 
   ## radioButtons
-  expect_equal(app$findWidget("radio")$setValue("1")$getValue(), "1")
-  expect_equal(app$findWidget("radio")$setValue("2")$getValue(), "2")
-  expect_equal(app$findWidget("radio")$setValue("3")$getValue(), "3")
+  expect_equal(roundtrip(app, "radio", "1"), "1")
+  expect_equal(roundtrip(app, "radio", "2"), "2")
+  expect_equal(roundtrip(app, "radio", "3"), "3")
 
   ## sliderInput, single
-  expect_equal(app$findWidget("slider1")$setValue(42)$getValue(), 42)
-  expect_equal(app$findWidget("slider1")$setValue(100)$getValue(), 100)
-  expect_equal(app$findWidget("slider1")$setValue(0)$getValue(), 0)
+  expect_equal(roundtrip(app, "slider1", 42), 42)
+  expect_equal(roundtrip(app, "slider1", 100), 100)
+  expect_equal(roundtrip(app, "slider1", 0), 0)
 
   ## sliderInput double
-  expect_equal(
-    app$findWidget("slider2")$setValue(c(42, 42))$getValue(),
-    c(42, 42)
-  )
-  expect_equal(
-    app$findWidget("slider2")$setValue(c(0, 100))$getValue(),
-    c(0,100)
-  )
-  expect_equal(
-    app$findWidget("slider2")$setValue(c(1, 4))$getValue(),
-    c(1, 4)
-  )
+  expect_equal(roundtrip(app, "slider2", c(42, 42)), c(42, 42))
+  expect_equal(roundtrip(app, "slider2", c(0, 100)), c(0,100))
+  expect_equal(roundtrip(app, "slider2", c(1, 4)), c(1, 4))
 
   ## textInput
-  expect_equal(
-    app$findWidget("text")$setValue("Hello world!")$getValue(),
-    "Hello world!"
-  )
+  expect_equal(roundtrip(app, "text", "Hello world!"), "Hello world!")
 
   ## passwordInput, TODO, this app does not have one
 })