+++ NEWS.md
@@ -4,6 +4,9 @@ shinytest 1.4.0.9002
 * `ShinyDriver$takeSnapshot()` gains ability to take a snapshot of a single
   element (#260).
 
+* New `ShinyDriver$waitForShiny()` that waits until Shiny is done computing
+  on the reactive graph (#327). 
+
 * `testApp()` can now take a path to a directory containing a single 
   interactive `.Rmd` (#334).